Improvements and new features for consideration
Feature:
Song details 'Extra Information' block
This gives typical a song details block on-screen which can be exported to png and pdf.  A higher quality PDF can be made by export to PNG and print to PDF using (for example) Xodo.
This gives a more 'printed' screen layout for those who like/expect it.  It needs to have a toggle of it's own (it is piggy backing on notes display).

Improvement:
2. Song lines are end trimmed on screen display (Users can add '_' rather than space to force a wide line).  This handles imported files with trailing spaces.