Feature: *** UPDATE - Title is now in bold rather than capitals.  This keep the width down which reduces the negative effect of a wide title causing autoscale to use smaller fonts.
Song details 'Extra Information' block
This gives typical a song details block on-screen which can be exported to png and pdf.  A higher quality PDF can be made by export to PNG and print to PDF using (for example) Xodo.
This gives a more 'printed' screen layout for those who like/expect it.  It needs to have a toggle of it's own (it is piggy backing on notes display).