Fixes: Autoscroll and pad start, nearby song sends, CCLI logging
Autsocroll - Fix for failed starts and incorrect stops of autoscroll (both stand alone and when using nearby)
- Now always cancels callbacks of autoscroll runnables immediatly before queue
- Increase of autoscroll delay from 2s to 8s when in host mode - this allows clients time to fully display songs and increases the rate of synchronised starts of host and clients.
- When in host mode the icon for the autoscoll timer is shown alone to indicate pending autoscroll start.

Pad start - This can be slow for in-built pads.  The pad icon is shown alone to indicate pending pad start.

Nearby song sends - Fix for significant stress and delays on slower clients as they try to keep up with rapid song changes.
-  For rapid song changes only the first and last song is sent.  The first is sent with no delay supporting 'normal' single song changes.  A subsequent rapid song change is delayed by 3s and can be cancelled by a further rapid song change. 3s after stabilising on a song this 'last' song is sent.

Testing of CCLI logging found
View CCLI log - not displaying some events
Fixes:
-  All calls to logging adjusted to send as many details as possible
-  A test for blank 'song' during build of report HTML has been changed to a test for blank 'action'
A combination of this 'song' test and not sending 'song' was stopping display of some log entries.
- The table build has been adjusted both to use names for colors and to correct a row format bug.