Menu xml adjustments - even 'flatter'! 1. Main menus - use of right arrow button icon with background colour the same as menu 2. Song menu - use of left arrow button with background colour same as menu, button repositioned to left
Search songs
1. Now displays 'folder/filename' instead of just folder, with this at the bottom. java and xml changes
2. No longer diaplys folders

For CCLI use the song title needs to be the authorised title.  Songs are ordered by filename in the 'Song menu', having filename visible during search is helpful.  For example, I use filenames of form '[Part of first line] ([Song title])' as it is common to refer to songs by first line - having the filename shown helps to select the song with the right [Part of the first line].